--- avalon/gui/delegates/__init__.py
@@ -1,22 +1,0 @@
-from ...vendor import Qt
-from Qt import QtCore, QtWidgets
-
-from ... import io
-
-from . import lib
-
-from .delegate_pretty_time import PrettyTimeDelegate
-from .delegate_version import VersionDelegate
-
-__all__ = [
-    "Qt",
-    "QtWidgets",
-    "QtCore",
-
-    "io",
-
-    "lib",
-
-    "PrettyTimeDelegate",
-    "VersionDelegate"
-]
