+++ avalon/gui/delegates/lib.py
@@ -2,7 +2,6 @@ import time
 from datetime import datetime
 import logging
 
-
 log = logging.getLogger(__name__)
