--- avalon/gui/models/__init__.py
@@ -1,51 +1,0 @@
-from ...vendor import Qt
-from Qt import QtCore, QtWidgets, QtGui
-from ...vendor import qtawesome
-from ... import io, api
-from ... import style
-
-from . import lib
-
-from .node import Node
-
-from .model_tree import TreeModel
-from .model_task import TaskModel
-from .model_asset import AssetModel
-from .model_subset import SubsetModel
-from .model_inventory import InventoryModel
-
-from .proxy_filter import FilterProxyModel
-from .proxy_exact_matches_filter import ExactMatchesFilterProxyModel
-from .proxy_recursive_sort_filter import RecursiveSortFilterProxyModel
-from .proxy_group_filter import GroupMemberFilterProxyModel
-from .proxy_subset_filter import SubsetFilterProxyModel
-from .proxy_family_filter import FamilyFilterProxyModel
-
-__all__ = [
-    "Qt",
-    "QtCore",
-    "QtWidgets",
-    "QtGui",
-
-    "qtawesome",
-    "io",
-    "api",
-    "style",
-
-    "lib",
-
-    "Node",
-
-    "TreeModel",
-    "TaskModel",
-    "AssetModel",
-    "SubsetModel",
-    "InventoryModel",
-
-    "FilterProxyModel",
-    "ExactMatchesFilterProxyModel",
-    "RecursiveSortFilterProxyModel",
-    "GroupMemberFilterProxyModel",
-    "SubsetFilterProxyModel",
-    "FamilyFilterProxyModel"
-]
