+++ avalon/gui/models/node.py
@@ -1,6 +1,5 @@
 import logging
 
-
 log = logging.getLogger(__name__)
