+++ avalon/gui/models/proxy_group_filter.py
@@ -1,4 +1,4 @@
-from . import QtCore
+from .. import QtCore
 from . import lib
