+++ avalon/gui/models/proxy_recursive_sort_filter.py
@@ -1,5 +1,5 @@
 import re
-from . import QtCore
+from .. import QtCore
 
 
 class RecursiveSortFilterProxyModel(QtCore.QSortFilterProxyModel):
