--- avalon/gui/views/__init__.py
@@ -1,16 +1,0 @@
-from ...vendor import Qt
-from Qt import QtCore, QtWidgets
-
-
-from .view_tree_deselectable import DeselectableTreeView
-from .view_assets import AssetsView
-
-
-__all__ = [
-    "Qt",
-    "QtCore",
-    "QtWidgets",
-
-    "DeselectableTreeView",
-    "AssetsView"
-]
