--- avalon/gui/widgets/__init__.py
@@ -1,38 +1,0 @@
-from ...vendor import Qt
-from Qt import QtCore, QtWidgets, QtGui
-from ...vendor import qtawesome
-from ... import io, api, pipeline, style
-
-
-from . import lib
-
-
-from .widget_subsets import SubsetsWidget
-from .widget_families_list import FamiliesListWidget
-from .widget_silos_tab import SilosTabWidget
-from .widget_assets import AssetsWidget
-from .widget_version_text_edit import VersionTextEdit
-from .widget_version import VersionWidget
-
-
-__all__ = [
-    "Qt",
-    "QtCore",
-    "QtWidgets",
-    "QtGui",
-
-    "qtawesome",
-    "io",
-    "api",
-    "pipeline",
-    "style",
-
-    "lib",
-
-    "SubsetsWidget",
-    "FamiliesListWidget",
-    "SilosTabWidget",
-    "AssetsWidget",
-    "VersionTextEdit",
-    "VersionWidget"
-]
