+++ avalon/tools/creator/app.py
@@ -5,7 +5,8 @@ import inspect
 from ...vendor.Qt import QtWidgets, QtCore, QtGui
 from ...vendor import qtawesome
 from ...vendor import six
-from ... import api, io, style
+from ... import api, io
+from ...gui import style
 from .. import lib
 
 module = sys.modules[__name__]
